refactor(cde): extract command logic into injected CdeApp
Program was a god class: composition root, CLI parser, and every command's
implementation, pulling dependencies ad-hoc via Resolve<T>() (service locator).

- Add CdeApp: command implementations with IFindService/ICatalogRepository/
  IMessageBus injected via constructor. Removes all 8 scattered Resolve<T>()
  calls; Program now resolves one CdeApp and dispatches parsed verbs to it
  (382 -> ~110 lines, single responsibility).
- Drop the Task.Run().Wait() sync-over-async wrappers around MessageBus.Send;
  call GetAwaiter().GetResult() directly (no threadpool hop).
- FindPopulous: remove the redundant second Where filter and the
  CompareDirEntries comparator; single Where + OrderByDescending.
- Extract MigrateOne, HandleReplCommand, PrintReplHelp from long methods.

Static Program.CreateCache/HashCatalog shims retained (delegate to CdeApp) so
cdeLibTest/DuplicationTest keeps driving scans through Program.
